+++ Sources/MapboxMaps/Style/LayerType.swift
@@ -37,8 +37,30 @@ public enum LayerType: String, Codable {
     /// Layer representing the sky
     case sky
 
-    /// Layer used for a 3D model
+    @available(*, deprecated, message: "Unsupported layer type")
     case model
+
+    public init?(rawValue: String) {
+        let supportedTypes: [LayerType] = [
+            .fill,
+            .line,
+            .symbol,
+            .circle,
+            .heatmap,
+            .fillExtrusion,
+            .raster,
+            .hillshade,
+            .background,
+            .locationIndicator,
+            .sky
+        ]
+
+        guard let matchingCase = supportedTypes.first(where: { $0.rawValue == rawValue }) else {
+            return nil
+        }
+
+        self = matchingCase
+    }
 
     /// The associated Swift struct type
     public var layerType: Layer.Type {
@@ -66,7 +88,7 @@ public enum LayerType: String, Codable {
         case .sky:
             return SkyLayer.self
         case .model:
-            return ModelLayer.self
+            fatalError("Not supported")
         }
     }
 }